Organize code into form_profile and customize style using bootstrap
